backend/requirements.txt
flask>=3.1.0
flask-cors>=5.0.0
pymilvus>=2.5.3
python-dotenv>=1.0.1
